Move logic to component folder
