--- frontend/src/Components/Tab.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-const Tab = (props) => {
-	if (props.isSelected) {
-		return <div>{props.children}</div>
-	}
-	return null
-}
-
-export default Tab
