+++ frontend/src/deadlineStatus/index.jsx
@@ -1,9 +1,11 @@
 import React from 'react'
+import TabSwitcher from './tabSwitcher'
 
 const DeadlineStatus = () => {
 	return (
 		<>
 			<div>Deadline Status</div>
+			<TabSwitcher />
 		</>
 	)
 }
