+++ frontend/src/deadlineStatus/tabSwitcher/Tab.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+const Tab = ({ isSelected, children }) => {
+	if (isSelected) {
+		return <div>{children}</div>
+	}
+	return null
+}
+
+export default Tab
